Basics of handing a dataset with python
The dataset is a time-series dataset, with weather conditions measured hourly. The first record is on 1/1/2012 at 0:00 and the last is on 12/31/2012 at 23:00. The dataset has a total of 8,784 observations and 7 parameters:

Temperature: Temp_C
Dew point Temperature: Dew Point Temp_C
Relative humidity: Rel Hum_%
Wind speed: Wind Speed_km/h
Visibility: Visibility_km
Pressure: Press_kPa
Weather

The basics of handling a dataset with python were applied by answering the following questions:
Q. 1)  Find all the unique 'Wind Speed' values in the data.
Q. 2) Find the number of times when the 'Weather is exactly Clear'.
Q. 3) Find the number of times when the 'Wind Speed was exactly 4 km/h'.
Q. 4) Find out all the Null Values in the data.
Q. 5) Rename the column name 'Weather' of the dataframe to 'Weather Condition'.
Q. 6) What is the mean 'Visibility' ?
Q. 7) What is the Standard Deviation of 'Pressure'  in this data?
Q. 8) What is the Variance of 'Relative Humidity' in this data?
Q. 9) Find all instances when 'Snow' was recorded.
Q. 10) Find all instances when 'Wind Speed is above 24' and 'Visibility is 25'.
Q. 11) What is the Mean value of each column against each 'Weather Condition?
Q. 12) What is the Minimum & Maximum value of each column against each 'Weather Condition?
Q. 13) Show all the Records where Weather Condition is Fog.
Q. 14) Find all instances when 'Weather is Clear' or 'Visibility is above 40'.
Q. 15) Find all instances when:'Weather is Clear' and 'Relative Humidity is greater than 50'